pyproject.toml
[tool.poetry]
name = "localize_be"
version = "0.1.0"
description = ""
authors = ["Nicolas Galler <ngaller@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
fake-useragent = "*"
beautifulsoup4 = "*"
requests = "*"
google-api-python-client = "*"
google-auth-httplib2 = "*"
google-auth-oauthlib = "*"
dagster = "^0.13.8"
pandas = "^1.3.4"
dagster-pandas = "^0.13.9"
scikit-learn = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
